Improvements after PR comments NB: We need to add more secrets before merging - Added more parameters to environment config - Moved article-query into article-component - Modified run-script to write new secrets to .env.production
